requirements.txt
motor
bcrypt
uvicorn
pymongo
fastapi
pydantic
passlib
python-dotenv
beanie
asyncio
gunicorn
python-keycloak